+++ pom.xml
@@ -3,13 +3,13 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.example</groupId>
-  <artifactId>demo</artifactId>
+  <groupId>de.ds</groupId>
+  <artifactId>keeptime</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <packaging>jar</packaging>
 
-  <name>demo</name>
-  <description>Demo project for Spring Boot</description>
+  <name>KeepTime</name>
+  <description>Time tracker</description>
 
   <parent>
     <groupId>org.springframework.boot</groupId>
