--- src/main/java/application/view/time/CallBackListener.java
@@ -1,5 +1,0 @@
-package application.view.time;
-
-public interface CallBackListener {
-	public void call();
-}
