+++ src/main/java/de/ds/keeptime/model/repos/ProjectRepository.java
@@ -1,9 +1,9 @@
-package application.model.repos;
+package de.ds.keeptime.model.repos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import application.model.Project;
+import de.ds.keeptime.model.Project;
 
 @Repository
 public interface ProjectRepository extends JpaRepository<Project, Long> {
