+++ src/main/java/de/ds/keeptime/model/repos/SettingsRepository.java
@@ -1,9 +1,9 @@
-package application.model.repos;
+package de.ds.keeptime.model.repos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import application.model.Settings;
+import de.ds.keeptime.model.Settings;
 
 @Repository
 public interface SettingsRepository extends JpaRepository<Settings, Long> {
