+++ src/main/java/de/ds/keeptime/model/repos/WorkRepository.java
@@ -1,4 +1,4 @@
-package application.model.repos;
+package de.ds.keeptime.model.repos;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import application.model.Work;
+import de.ds.keeptime.model.Work;
 
 @Repository
 public interface WorkRepository extends JpaRepository<Work, Long> {
