+++ src/main/java/de/ds/keeptime/view/time/CallBackListener.java
@@ -1,0 +1,5 @@
+package de.ds.keeptime.view.time;
+
+public interface CallBackListener {
+	public void call();
+}
